third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/css-module/css-module-worker-test-expected.txt

This is a testharness.js-based test.
[FAIL] A dynamic import CSS Module within a web worker should not load and should not attempt to fetch the module.
  assert_equals: Shouldn't have tried fetching CSS module in worker expected 0 but got 1
Harness: the test ran to completion.
